requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
PyJWT
praw